2. Social-Media-Dashboard-using-Django-main/2. Social-Media-Dashboard-using-Django-main/Social-Media-Dashboard-using-Django-main/requirements.txt
Django~=4.0
django-crispy-forms~=1.14
django-allauth~=0.51
tweepy~=4.12
facebook-sdk~=3.1 # Or another Facebook library
python-decouple # For environment variables